+++ tests/Biblys/Legacy/OrderDeliveryHelpersTest.php
@@ -17,6 +17,7 @@ use Entity\Exception\CartException;
 use Exception;
 use Mockery;
 use Model\Article;
+use Model\Cart;
 use OrderManager;
 use PHPUnit\Framework\TestCase;
 use Propel\Runtime\Exception\PropelException;
@@ -50,8 +51,11 @@ class OrderDeliveryHelpersTest extends TestCase
             ->with("order_phone_required")
             ->andReturn(0);
 
+        $cart = Mockery::mock(Cart::class);
+        $cart->shouldReceive("containsDownloadableArticles")->andReturn(false);
+
         // when
-        OrderDeliveryHelpers::validateOrderDetails($request, $currentSite);
+        OrderDeliveryHelpers::validateOrderDetails($request, $currentSite, $cart);
 
         // then
         $this->expectNotToPerformAssertions();
@@ -80,12 +84,15 @@ class OrderDeliveryHelpersTest extends TestCase
             ->with("order_phone_required")
             ->andReturn(1);
 
+        $cart = Mockery::mock(Cart::class);
+        $cart->shouldReceive("containsDownloadableArticles")->andReturn(false);
+
         // then
         $this->expectException(OrderDetailsValidationException::class);
         $this->expectExceptionMessage("Le champ &laquo;&nbsp;Numéro de téléphone&nbsp;&raquo; est obligatoire.");
 
         // when
-        OrderDeliveryHelpers::validateOrderDetails($request, $currentSite);
+        OrderDeliveryHelpers::validateOrderDetails($request, $currentSite, $cart);
     }
 
     /**
@@ -111,8 +118,82 @@ class OrderDeliveryHelpersTest extends TestCase
             ->with("order_phone_required")
             ->andReturn(1);
 
+        $cart = Mockery::mock(Cart::class);
+        $cart->shouldReceive("containsDownloadableArticles")->andReturn(false);
+
         // when
-        OrderDeliveryHelpers::validateOrderDetails($request, $currentSite);
+        OrderDeliveryHelpers::validateOrderDetails($request, $currentSite, $cart);
+
+        // then
+        $this->expectNotToPerformAssertions();
+    }
+
+    /**
+     * @throws OrderDetailsValidationException
+     * @throws Exception
+     */
+    public function testValidateOrderDetailsThrowsWhenCartContainsDownloadableAndCheckboxIsUnchecked()
+    {
+        // given
+        $request = new Request();
+        $request->request->set("order_firstname", "Victor");
+        $request->request->set("order_lastname", "Hugo");
+        $request->request->set("order_address1", "Place des Vosges");
+        $request->request->set("order_postalcode", "75004");
+        $request->request->set("order_city", "Paris");
+        $request->request->set("country_id", 1);
+        $request->request->set("order_email", "victor.hugo@biblys.fr");
+        $request->request->set("cgv_checkbox", 1);
+        $request->request->set("order_phone", "06");
+
+        $currentSite = Mockery::mock(CurrentSite::class);
+        $currentSite->shouldReceive("getOption")
+            ->with("order_phone_required")
+            ->andReturn(1);
+
+        $cart = Mockery::mock(Cart::class);
+        $cart->shouldReceive("containsDownloadableArticles")->andReturn(true);
+
+        // then
+        $this->expectException(OrderDetailsValidationException::class);
+        $this->expectExceptionMessage(
+            "Vous devez accepter les conditions spécifiques au numérique, " .
+            "car votre panier contient des articles téléchargeables."
+        );
+
+        // when
+        OrderDeliveryHelpers::validateOrderDetails($request, $currentSite, $cart);
+    }
+
+    /**
+     * @throws OrderDetailsValidationException
+     * @throws Exception
+     */
+    public function testValidateOrderDetailsThrowsWhenCartContainsDownloadableAndCheckboxIsChecked()
+    {
+        // given
+        $request = new Request();
+        $request->request->set("order_firstname", "Victor");
+        $request->request->set("order_lastname", "Hugo");
+        $request->request->set("order_address1", "Place des Vosges");
+        $request->request->set("order_postalcode", "75004");
+        $request->request->set("order_city", "Paris");
+        $request->request->set("country_id", 1);
+        $request->request->set("order_email", "victor.hugo@biblys.fr");
+        $request->request->set("order_phone", "01234567890");
+        $request->request->set("cgv_checkbox", 1);
+        $request->request->set("downloadable_articles_checkbox", "1");
+
+        $currentSite = Mockery::mock(CurrentSite::class);
+        $currentSite->shouldReceive("getOption")
+            ->with("order_phone_required")
+            ->andReturn(1);
+
+        $cart = Mockery::mock(Cart::class);
+        $cart->shouldReceive("containsDownloadableArticles")->andReturn(true);
+
+        // when
+        OrderDeliveryHelpers::validateOrderDetails($request, $currentSite, $cart);
 
         // then
         $this->expectNotToPerformAssertions();